requirements.txt
pandas>=2.0
pyarrow>=14.0
requests>=2.31
prefect>=2.14
prefect-gcp>=0.5
google-cloud-bigquery>=3.13
google-cloud-storage>=2.13
streamlit>=1.30
db-dtypes>=1.2
dbt-core>=1.7
dbt-bigquery>=1.7
